refactor: move episodes outside of series module
