+++ src/episodes/helpers/scope.ts
