+++ src/routes.ts
@@ -3,7 +3,7 @@ import { AppContext } from './common/types/core.ts';
 import { config } from './config/index.ts';
 import { news, newsWorker } from './news/index.ts';
 import { series } from './series/index.ts';
-import { seriesEpisodes } from './series/episodes/episodes.controller.ts';
+import { episodes } from './episodes/episodes.controller.ts';
 
 const router = new Router({
   methods: ['GET'],
@@ -27,8 +27,8 @@ router.get('/series', async (ctx: AppContext) => {
   await series(ctx);
 });
 
-router.get('/series/episodes', async (ctx: AppContext) => {
-  await seriesEpisodes(ctx);
+router.get('/episodes', async (ctx: AppContext) => {
+  await episodes(ctx);
 });
 
 export default router;
